Switch from double to single precission floats

+++ src/acceleration/aabb.hpp
@@ -15,14 +15,14 @@ struct Aabb {
     /// @param tMax Maximum distance
     /// @param rec HitRecord, for which only t is recorded yet.
     /// @return True if there is a hit.
-    bool hit(const Ray& ray, double tMin, double tMax, HitRecord& rec) const;
+    bool hit(const Ray& ray, float tMin, float tMax, HitRecord& rec) const;
 
     /// @brief Simplest intersect function. Only returns true if there is an
     /// intersection. Kept for legacy.
     /// @param ray Intersecting ray
     /// @param tMin Minimum distance
     /// @param tMax Maximum distance
-    bool intersect(const Ray& ray, const double tMin, const double tMax) const;
+    bool intersect(const Ray& ray, const float tMin, const float tMax) const;
 
     /// @brief Grow AABB to fit new point `pt`.
     void grow(Vec3 pt);
